+++ packages/lib/test/actionMiddleware/onActionMiddleware.test.ts
@@ -1,16 +1,14 @@
-import { isObservable, observable } from "mobx"
+import { observable } from "mobx"
 import {
   ActionCall,
   ActionContext,
   applySnapshot,
-  deserializeActionCallArgument,
   getSnapshot,
   model,
   Model,
   modelAction,
   onActionMiddleware,
   prop,
-  serializeActionCallArgument,
 } from "../../src"
 import "../commonSetup"
 import { autoDispose } from "../utils"
@@ -76,6 +74,7 @@ test("onActionMiddleware", () => {
             1,
           ],
           "targetPath": Array [],
+          "targetPathIds": Array [],
         },
         Object {
           "actionName": "addX",
@@ -126,6 +125,9 @@ test("onActionMiddleware", () => {
           "targetPath": Array [
             "p2",
           ],
+          "targetPathIds": Array [
+            "id-2",
+          ],
         },
         Object {
           "actionName": "addY",
@@ -168,6 +170,7 @@ test("onActionMiddleware", () => {
             4,
           ],
           "targetPath": Array [],
+          "targetPathIds": Array [],
         },
         Object {
           "actionName": "addXY",
@@ -220,6 +223,7 @@ test("onActionMiddleware", () => {
             RandomClass {},
           ],
           "targetPath": Array [],
+          "targetPathIds": Array [],
         },
         Object {
           "actionName": "other",
@@ -277,6 +281,7 @@ test("onActionMiddleware", () => {
             ],
           ],
           "targetPath": Array [],
+          "targetPathIds": Array [],
         },
         Object {
           "actionName": "other",
@@ -339,6 +344,7 @@ test("onActionMiddleware", () => {
             },
           ],
           "targetPath": Array [],
+          "targetPathIds": Array [],
         },
         Object {
           "actionName": "other",
@@ -404,6 +410,9 @@ test("onActionMiddleware", () => {
           "targetPath": Array [
             "p2",
           ],
+          "targetPathIds": Array [
+            "id-2",
+          ],
         },
         Object {
           "actionName": "$$applySnapshot",
@@ -442,73 +451,4 @@ test("onActionMiddleware", () => {
   p1.addXY(5, 6)
   p2.addXY(5, 6)
   expect(events).toMatchInlineSnapshot(`Array []`)
-})
-
-test("serializeActionCallArgument and deserializeActionCallArgument", () => {
-  // unserializable args
-  class RandomClass {}
-  const rc = new RandomClass()
-
-  expect(() => serializeActionCallArgument(rc)).toThrow(
-    "serializeActionCallArgument could not serialize the given value"
-  )
-
-  // primitive
-  expect(serializeActionCallArgument(42)).toBe(42)
-  expect(deserializeActionCallArgument(42)).toBe(42)
-
-  // date
-  expect(serializeActionCallArgument(new Date(1000))).toEqual({ $dateAsTimestamp: 1000 })
-  expect(deserializeActionCallArgument({ $dateAsTimestamp: 1000 })).toEqual(new Date(1000))
-
-  // plain obj
-  const obj = { x: 10 }
-
-  expect(serializeActionCallArgument(obj)).toEqual(obj)
-  expect(deserializeActionCallArgument(obj)).toEqual(obj)
-
-  // observable obj
-  const obsObj = observable(obj)
-
-  expect(serializeActionCallArgument(obsObj)).toEqual(obj)
-  expect(isObservable(serializeActionCallArgument(obsObj))).toBe(false)
-
-  // array
-  const arr = [{ x: 10 }, 20]
-
-  expect(serializeActionCallArgument(arr)).toEqual(arr)
-  expect(deserializeActionCallArgument(arr)).toEqual(arr)
-
-  // observable array
-  const obsArr = observable(arr)
-
-  expect(serializeActionCallArgument(obsArr)).toEqual(arr)
-  expect(isObservable(serializeActionCallArgument(obsArr))).toBe(false)
-  expect(isObservable(serializeActionCallArgument(obsArr)[0])).toBe(false)
-
-  // map
-  const mapKV: [any, any][] = [["x", 10], ["y", { z: 20 }]]
-  const map = new Map<any, any>(mapKV)
-
-  expect(serializeActionCallArgument(map)).toEqual({
-    $mapAsArray: mapKV,
-  })
-  const mapBack = deserializeActionCallArgument({
-    $mapAsArray: mapKV,
-  })
-  expect(mapBack instanceof Map).toBe(true)
-  expect(Array.from(mapBack.entries())).toEqual(mapKV)
-
-  // set
-  const setK: any[] = ["x", { z: 20 }]
-  const set = new Set<any>(setK)
-
-  expect(serializeActionCallArgument(set)).toEqual({
-    $setAsArray: setK,
-  })
-  const setBack = deserializeActionCallArgument({
-    $setAsArray: setK,
-  })
-  expect(setBack instanceof Set).toBe(true)
-  expect(Array.from(setBack.keys())).toEqual(setK)
 })