+++ packages/site/src/examples/clientServer/server.ts
@@ -18,14 +18,13 @@ class Server {
   sendMessage(actionCall: ActionCall) {
     // the timeouts are just to simulate network delays
     setTimeout(() => {
-      const deserializedActionCall = deserializeActionCall(actionCall)
-
       // apply the action over the server root store
       // sometimes applying actions might fail (for example on invalid operations
       // such as when one client asks to delete a model from an array and other asks to mutate it)
       // so we try / catch it
       let applyActionSucceeded = false
       try {
+        const deserializedActionCall = deserializeActionCall(this.serverRootStore, actionCall)
         applyAction(this.serverRootStore, deserializedActionCall)
         applyActionSucceeded = true
       } catch (err) {